+++ commands/repeat.js
@@ -1,11 +1,24 @@
+const Discord = require('discord.js')
+
 module.exports = {
   name: 'repeat',
   aliases: ['loop', 'rp'],
   inVoiceChannel: true,
   run: async (client, message, args) => {
     const queue = client.distube.getQueue(message)
-    if (!queue) return message.channel.send(`${client.emotes.error} | There is nothing playing!`)
-    let mode = null
+    if (!queue) {
+      return message.reply({
+        embeds: [
+          new Discord.EmbedBuilder()
+            .setColor('#ec9bbb')
+            .setTitle(`${client.emotes.error} | Error`)
+            .setDescription('There is no song in the current queue!')
+            .setTimestamp()
+            .setFooter({ text: `${client.user.tag}`, iconURL: '<put your bot avatar link>' })
+        ]
+      })
+    }
+    let mode
     switch (args[0]) {
       case 'off':
         mode = 0
@@ -18,7 +31,15 @@ module.exports = {
         break
     }
     mode = queue.setRepeatMode(mode)
-    mode = mode ? (mode === 2 ? 'Repeat queue' : 'Repeat song') : 'Off'
-    message.channel.send(`${client.emotes.repeat} | Set repeat mode to \`${mode}\``)
+    mode = mode ? (mode === 2 ? 'Queue' : 'Current Song') : 'Off'
+    message.reply({
+      embeds: [
+        new Discord.EmbedBuilder()
+          .setColor('#ec9bbb')
+          .setDescription(`Repeat Mode has been set to \`${mode}\`.`)
+          .setTimestamp()
+          .setFooter({ text: `${client.user.tag}`, iconURL: '<put your bot avatar link>' })
+      ]
+    })
   }
 }