+++ commands/stop.js
@@ -2,7 +2,6 @@ const Discord = require('discord.js')
 
 module.exports = {
   name: 'stop',
-  aliases: ['disconnect', 'leave'],
   inVoiceChannel: true,
   run: async (client, message) => {
     const queue = client.distube.getQueue(message)